[CHK-702] fix: update currency format, optimize imports

+++ src/main/java/it/pagopa/ecommerce/payment/methods/domain/valueobjects/PspAmount.java
@@ -4,20 +4,21 @@ import lombok.EqualsAndHashCode;
 import org.springframework.lang.NonNull;
 
 import java.io.Serializable;
+import java.math.BigInteger;
 import java.util.Objects;
 
 @ValueObjects
 @EqualsAndHashCode
 public class PspAmount implements Serializable {
 
-    private final Double amount;
+    private final BigInteger amount;
 
-    public PspAmount(@NonNull Double amount) {
+    public PspAmount(@NonNull BigInteger amount) {
 
         this.amount = Objects.requireNonNull(amount);
     }
 
-    public @NonNull Double value() {
+    public @NonNull BigInteger value() {
         return amount;
     }
 }
